A shapshot of my current Fortranification progress
The goal is to make the Hdf module support loading data
in both c_layout and fortran_layout Bigarrays.  At this
point, I think the best way to do this is by making the
module in to a functor which can be instantiated with
either the c_layout or fortran_layout option.  The first-
class module support in OCaml 3.12 may offer some options
on this front as well.