Finish last commit's changes; Use Maps instead of Arrays for collections
Attributes, Vdata Fields, *.read_all all work with maps now instead of
arrays.  This makes it easier to locate specific data without having to
know an offset.

That said - the writing functions may need to be altered in order to order
output data properly, particularly in the case of Vdata.  THIS IS UNTESTED!